Move processors.rs contents to mod.rs; no need to be separate.
